+++ generator/test/generator_test.dart
@@ -226,46 +226,6 @@ class _Bad {
     );
   });
 
-  test('primary key not final', () async {
-    await expectLater(
-      () async => await testBuilder(
-        generateRealmObjects(),
-        {
-          'pkg|lib/src/test.dart': r'''
-import 'package:realm_common/realm_common.dart';
-
-part 'test.g.dart';
-
-@RealmModel()
-class _Bad {
-  @PrimaryKey()
-  late int primaryKeyIsNotFinal;
-}'''
-        },
-        reader: await PackageAssetReader.currentIsolate(),
-      ),
-      throwsA(
-        isA<RealmInvalidGenerationSourceError>().having(
-          (e) => e.format(),
-          'format()',
-          'Primary key field is not final\n'
-              '\n'
-              'in: package:pkg/src/test.dart:8:12\n'
-              '  ╷\n'
-              '5 │ @RealmModel()\n'
-              '6 │ class _Bad {\n'
-              '  │       ━━━━ in realm model for \'Bad\'\n'
-              '7 │   @PrimaryKey()\n'
-              '8 │   late int primaryKeyIsNotFinal;\n'
-              '  │            ^^^^^^^^^^^^^^^^^^^^ is not final\n'
-              '  ╵\n'
-              'Add a final keyword to the definition of \'primaryKeyIsNotFinal\', or remove the @PrimaryKey annotation.\n'
-              '',
-        ),
-      ),
-    );
-  });
-
   test('primary keys always indexed', () async {
     final sb = StringBuffer();
     var done = false;