+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+vNext
+==============================================================
+
+### Changes
+* Primary key annotation no longer requires field to be final.
+
+
 0.2.0+alpha Release notes (2022-01-31)
 ==============================================================